Deal with vertical margins and fractions of heights
I've added two extensions....

Deal with vertical margins:
By calculating the height of an object with .outerHeight(true) and setting the new height the same way it's possible to include margins in the calculation. This makes especially sense in combination with the other extension.

Deal with fractions of the height:
The idea is that if you have some kind of a setup with boxes where some boxes must be exactly half the height of the other boxes you can specify the fraction with "vertilize='0.5'". If you leave it out (just use "vertilize") it defaults to 1 which leads to the default behaviour. It allows to have any fraction, i.e. half, third, quarter, ...

If you have equal margins on the full and half boxes in such a setup then the combination of the two leads to a layout where you have the backgrounds horizontally aligned.